backend/requirements.txt
flask
flask-cors
tensorflow
pillow
matplotlib
numpy
python-dotenv
